Add run-time dependencies
The run-time dependencies when pulled in through NPM do not match the run-time dependencies listed in the Bower package. This means that when using NPM for dependency management the package's dependencies will not be pulled in. This update simply makes the dependencies in package.json match those of bower.json.